+++ src/config.js
@@ -1,9 +1,9 @@
 /*
-* src/config.js
+* src/config.js     <-- tbd. rename 'emulationParams.js'???
 *
-* Provides access to e.g. 'firebase.json'
+* Provides access to emulator configuration
 */
-import fs from "fs";
+import fs from "fs"
 
 const fn = process.env["FIREBASE_JSON"] || './firebase.json';
 
@@ -11,6 +11,21 @@ const firebaseJson = JSON.parse(
     fs.readFileSync(fn, 'utf8')
 );
 
+const FIRESTORE_HOST = process.env["FIRESTORE_EMULATOR_HOST"] || (() => {
+  const port = firebaseJson.emulators.firestore.port;   // "6767"
+  if (!port) {
+    throw new Error(`Unable to get Firestore emulator port from '${fn}'`);
+  }
+
+  return `localhost:${port}`;
+})();
+
+// This is constant - there is no way to either set it, or sniff it out? (firebase 8.8.1)
+//
+const FUNCTIONS_URL = "http://localhost:5001";    // not available in any Firebase config, to read. :(
+
 export {
-  firebaseJson
+  firebaseJson,
+  FIRESTORE_HOST,
+  FUNCTIONS_URL
 }