+++ src/firestoreTesting/readOnly.js
@@ -11,7 +11,9 @@
 */
 import { strict as assert } from 'assert'
 
-import firebase from '@firebase/testing'
+import firebase from '@firebase/rules-unit-testing'
+import admin from 'firebase-admin'
+
 import { projectId } from './common.js'
 import { emul } from './emul.js'
 
@@ -21,14 +23,22 @@ import { PRIME_ROUND } from "./common"
 assert(!PRIME_ROUND);
 
 /*
+* Create an "admin" Firestore instance, for restoring the values.
+*
+* Note: We could use '.initializeAdminApp' but since it's just a veneer above 'firebase-admin', we use that straight.
+*/
+const adminApp = admin.initializeApp({
+  projectId
+}, "rules-test-admin" );    // keep away from other apps; not needed to be random since initialized only once
+
+/*
 * Unlike in the Firestore API, we allow authentication to be set after collection.
 */
 const dbAuth = {  // firebase.firestore.Firestore -like
 
   collection: collectionPath => {   // string => { as: ... }
-    const collAdmin = firebase.initializeAdminApp({   // same for all auths
-      projectId
-    }).firestore().collection(collectionPath);
+    const collAdmin = adminApp.firestore()
+        .collection(collectionPath);
 
     return {
       as: auth => {   // { uid: string }|null => firebase.firestore.CollectionReference -like
@@ -47,7 +57,9 @@ const dbAuth = {  // firebase.firestore.Firestore -like
 * We know how to clean up after ourselves. Runs after all the tests (once per each suite).
 */
 afterAll( async () => {   // () => Promise of ()
-  await Promise.all( firebase.apps().map( app => { return app.delete(); }));
+  const apps = [ adminApp, ...firebase.apps() ];
+
+  await Promise.all( apps.map( app => { return app.delete(); }));
 });
 
 // Enable '.toAllow' and '.toDeny' matchers, as a side effect